refactor(structure): split spmc queue base header
